Backend/requirements.txt
langchain==0.3.21
langgraph==0.3.11
language-tool-python==2.6.5
langchain_openai==0.3.9
langchain_community==0.3.20
gtts
python-dotenv==1.0.1
fastapi==0.115.11
uvicorn==0.34.0
googletrans

#pip install crewai crewai_tools langchain langchain_community langchain_openai